+++ tests/testthat/test_MeasureCiNestedCV.R
@@ -2,7 +2,7 @@ test_that("basic", {
   task = tsk("mtcars")$cbind(data.frame(chas = rep(c("a", "b"), times = 16)))
   expect_ci_measure(
     "ci.ncv",
-    rsmp("nested_cv", folds = 3L, repeats = 5L),
+    rsmp("ncv", folds = 3L, repeats = 5L),
     task = task
   )
 })
@@ -10,7 +10,7 @@ test_that("basic", {
 test_that("aggr and CI point estimate agree", {
   withr::local_seed(1)
   task = tsk("iris")
-  rr = resample(task, lrn("classif.featureless"), rsmp("nested_cv", folds = 3L, repeats = 5L))
+  rr = resample(task, lrn("classif.featureless"), rsmp("ncv", folds = 3L, repeats = 5L))
   ci = rr$aggregate(msr("ci", "classif.acc"))
   aggr = rr$aggregate(msr("classif.acc"))
   # there is some difference due to how we do the aggregation.